convert discussion row menu
