+++ client/styles/pages/discussions/discussion_row_menu.scss
@@ -1,25 +1,28 @@
 @import 'client/styles/shared';
 
 .DiscussionRowMenu {
-    display: inline-block;
-    position: relative;
-    left: 9px;
-    border-radius: 9999px;
-    width: 20px;
-    height: 20px;
-    position: relative;
-    top: 1px;
-    text-align: center;
-    svg {
-        cursor: pointer;
-        stroke: $text-color-light;
-        position: relative;
-        top: 2px;
-    }
-    &:hover {
-        background: $background-color-light;
-        svg {
-            stroke: $text-color-medium;
-        }
-    }
+  //   border-radius: 9999px;
+  //   display: inline-block;
+  //   height: 20px;
+  //   left: 9px;
+  //   position: relative;
+  //   position: relative;
+  //   text-align: center;
+  //   top: 1px;
+  //   width: 20px;
+
+  //   svg {
+  //     cursor: pointer;
+  //     position: relative;
+  //     stroke: $text-color-light;
+  //     top: 2px;
+  //   }
+
+  //   &:hover {
+  //     background: $background-color-light;
+
+  //     svg {
+  //       stroke: $text-color-medium;
+  //     }
+  //   }
 }
